+++ psycopg/psycopg/postgres.py
@@ -119,7 +119,7 @@ TEXT_ARRAY_OID = types["text"].array_oid
 
 def register_default_adapters(context: AdaptContext) -> None:
 
-    from .types import array, bool, composite, datetime, json
+    from .types import array, bool, composite, datetime, json, multirange
     from .types import net, none, numeric, range, string, uuid
 
     array.register_default_adapters(context)
@@ -127,6 +127,7 @@ def register_default_adapters(context: AdaptContext) -> None:
     composite.register_default_adapters(context)
     datetime.register_default_adapters(context)
     json.register_default_adapters(context)
+    multirange.register_default_adapters(context)
     net.register_default_adapters(context)
     none.register_default_adapters(context)
     numeric.register_default_adapters(context)